requirements.txt
asttokens==2.4.1
colorama==0.4.6
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.2
openslide-python==1.3.1
packaging==23.2
pandas==2.1.3
parso==0.8.3
Pillow==10.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
pydicom==2.4.3
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.2
setuptools==68.2.2
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.0
tzdata==2023.3
wcwidth==0.2.12
